Create eslint and prettier configuration
This uses pre-commit to run the linter and code formatting via eslint
and prettier.

The eslint configuration is designed to work with the plain html and JS
modules that themes use (no webpack to resolve imports).

We're using pre-commit instead of husky to run the git hooks in case we
need to use Python to create repository automations.