Add Light color scheme
These values are from the Rubin Visual Identity Manual (with the
exception of the pink for "love", which is held over from the Discourse
default theme.

Note that this schema implements a dark header, which I think will be
useful for letting us use the same logo for both light and dark themes.